+++ packages/interwovenkit-react/src/pages/bridge/BridgeHistory.module.css
@@ -13,18 +13,3 @@
 .item {
   border-bottom: 1px solid var(--border);
 }
-
-.showMore {
-  color: var(--gray-3);
-  margin: 16px auto;
-  font-size: 12px;
-  font-weight: 500;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 2px;
-
-  &:hover {
-    color: var(--gray-1);
-  }
-}
