+++ packages/interwovenkit-react/src/pages/bridge/BridgeHistory.tsx
@@ -1,13 +1,13 @@
 import { useState } from "react"
 import { useToggle } from "react-use"
-import { IconChevronDown } from "@initia/icons-react"
 import { useInterwovenKit } from "@/public/data/hooks"
 import Page from "@/components/Page"
 import Status from "@/components/Status"
 import AsyncBoundary from "@/components/AsyncBoundary"
 import CheckboxButton from "@/components/CheckboxButton"
+import LoadMoreButton from "@/components/LoadMoreButton"
 import {
-  BRIDGE_HISTORY_ITEMS_PER_PAGE,
+  BRIDGE_HISTORY_LIMIT_PER_PAGE,
   BRIDGE_HISTORY_LIMIT,
   useBridgeHistoryList,
 } from "./data/history"
@@ -29,7 +29,7 @@ const BridgeHistory = () => {
   const [page, setPage] = useState(1)
   const [showAll, toggleShowAll] = useToggle(!myHistory.length)
   const filteredHistory = showAll ? allHistory : myHistory
-  const paginatedHistory = filteredHistory.slice(0, page * BRIDGE_HISTORY_ITEMS_PER_PAGE)
+  const paginatedHistory = filteredHistory.slice(0, page * BRIDGE_HISTORY_LIMIT_PER_PAGE)
 
   return (
     <Page title="Bridge/Swap activity">
@@ -57,10 +57,8 @@ const BridgeHistory = () => {
           ))
         )}
 
-        {filteredHistory.length > page * BRIDGE_HISTORY_ITEMS_PER_PAGE ? (
-          <button onClick={() => setPage((page) => page + 1)} className={styles.showMore}>
-            Show More <IconChevronDown size={14} />
-          </button>
+        {filteredHistory.length > page * BRIDGE_HISTORY_LIMIT_PER_PAGE ? (
+          <LoadMoreButton onClick={() => setPage((page) => page + 1)} />
         ) : (
           history.length >= BRIDGE_HISTORY_LIMIT && (
             <Status>
