+++ packages/interwovenkit-react/src/pages/wallet/tabs/nft/CollectionDetails.tsx
@@ -2,7 +2,7 @@ import { useLocationState, useNavigate } from "@/lib/router"
 import type { NormalizedChain } from "@/data/chains"
 import { parsePaginatedResponse } from "@/data/pagination"
 import Page from "@/components/Page"
-import LoadMoreButton from "../../components/LoadMoreButton"
+import LoadMoreButton from "@/components/LoadMoreButton"
 import type { NormalizedCollection } from "./queries"
 import { useNfts } from "./queries"
 import WithNormalizedNft from "./WithNormalizedNft"
